Added component logic hook useProjectPhases
