+++ SharePointFramework/ProjectWebParts/src/components/ProjectPhases/getPhaseSitePages.ts
@@ -13,15 +13,16 @@ export const getPhaseSitePages = async (phases: ProjectPhaseModel[]) => {
       .getByTitle('Områdesider')
       .items.select('Id, Title, FileRef, EncodedAbsUrl, FileLeafRef')
       .get()
+
     sitePages = sitePages.filter((p) => {
       return phases.some((phase) => phase.name === p.Title)
     })
 
-    const phaseSitePages: IPhaseSitePageModel[] = sitePages.map((p) => ({
+    const phaseSitePages = sitePages.map((p) => ({
       id: p.Id,
       title: p.Title,
       fileLeafRef: p.FileLeafRef
-    }))
+    }) as IPhaseSitePageModel)
 
     return phaseSitePages
   } catch (error) {
